Refactor CID helpers into dedicated module
