+++ cid_utils.py
@@ -1,0 +1,377 @@
+import base64
+import hashlib
+import json
+from urllib.parse import urlparse
+
+import requests
+from flask import make_response, request
+
+try:
+    from db_access import (
+        create_cid_record,
+        get_cid_by_path,
+        get_user_servers,
+        get_user_variables,
+        get_user_secrets,
+    )
+except (RuntimeError, ImportError):
+    create_cid_record = None
+    get_cid_by_path = None
+    get_user_servers = None
+    get_user_variables = None
+    get_user_secrets = None
+
+
+def _ensure_db_access():
+    global create_cid_record, get_cid_by_path, get_user_servers, get_user_variables, get_user_secrets
+    if None in (create_cid_record, get_cid_by_path, get_user_servers, get_user_variables, get_user_secrets):
+        from db_access import (
+            create_cid_record as _create_cid_record,
+            get_cid_by_path as _get_cid_by_path,
+            get_user_servers as _get_user_servers,
+            get_user_variables as _get_user_variables,
+            get_user_secrets as _get_user_secrets,
+        )
+
+        if create_cid_record is None:
+            create_cid_record = _create_cid_record
+        if get_cid_by_path is None:
+            get_cid_by_path = _get_cid_by_path
+        if get_user_servers is None:
+            get_user_servers = _get_user_servers
+        if get_user_variables is None:
+            get_user_variables = _get_user_variables
+        if get_user_secrets is None:
+            get_user_secrets = _get_user_secrets
+
+# ============================================================================
+# CID GENERATION AND STORAGE HELPERS
+# ============================================================================
+
+def generate_cid(file_data):
+    """Generate a simple CID-like hash from file data only"""
+    hasher = hashlib.sha256()
+    hasher.update(file_data)
+    sha256_hash = hasher.digest()
+
+    encoded = base64.b32encode(sha256_hash).decode('ascii').lower().rstrip('=')
+    return f"bafybei{encoded[:52]}"  # Truncate to reasonable length
+
+
+def process_file_upload(form):
+    """Process file upload from form and return file content and filename"""
+    uploaded_file = form.file.data
+    file_content = uploaded_file.read()
+    filename = uploaded_file.filename or 'upload'
+    return file_content, filename
+
+
+def process_text_upload(form):
+    """Process text upload from form and return file content"""
+    text_content = form.text_content.data
+    file_content = text_content.encode('utf-8')
+    return file_content
+
+
+def process_url_upload(form):
+    """Process URL upload from form by downloading content and return file content and MIME type"""
+    url = form.url.data.strip()
+
+    try:
+        headers = {
+            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 '
+                          '(KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
+        }
+
+        response = requests.get(url, timeout=30, headers=headers, stream=True)
+        response.raise_for_status()
+
+        content_type = response.headers.get('content-type', 'application/octet-stream')
+        mime_type = content_type.split(';')[0].strip().lower()
+
+        content_length = response.headers.get('content-length')
+        if content_length and int(content_length) > 100 * 1024 * 1024:
+            raise ValueError("File too large (>100MB)")
+
+        file_content = b''
+        downloaded_size = 0
+        max_size = 100 * 1024 * 1024
+
+        for chunk in response.iter_content(chunk_size=8192):
+            if chunk:
+                downloaded_size += len(chunk)
+                if downloaded_size > max_size:
+                    raise ValueError("File too large (>100MB)")
+                file_content += chunk
+
+        parsed_url = urlparse(url)
+        filename = parsed_url.path.split('/')[-1]
+
+        if not filename or '.' not in filename:
+            extension = get_extension_from_mime_type(mime_type)
+            if extension:
+                filename = f"download{extension}"
+            else:
+                filename = "download"
+
+        return file_content, mime_type
+
+    except requests.exceptions.RequestException as e:
+        raise ValueError(f"Failed to download from URL: {str(e)}")
+    except Exception as e:
+        raise ValueError(f"Error processing URL: {str(e)}")
+
+
+def save_server_definition_as_cid(definition, user_id):
+    """Save server definition as CID and return the CID string"""
+    _ensure_db_access()
+    definition_bytes = definition.encode('utf-8')
+    cid = generate_cid(definition_bytes)
+
+    existing_cid = get_cid_by_path(f"/{cid}")
+    if not existing_cid:
+        create_cid_record(cid, definition_bytes, user_id)
+
+    return cid
+
+
+def store_cid_from_json(json_content, user_id):
+    """Store JSON content in a CID record and return the CID"""
+    _ensure_db_access()
+    json_bytes = json_content.encode('utf-8')
+    cid = generate_cid(json_bytes)
+
+    existing_cid = get_cid_by_path(f"/{cid}")
+    if not existing_cid:
+        create_cid_record(cid, json_bytes, user_id)
+
+    return cid
+
+
+# ============================================================================
+# DEFINITIONS CID HELPERS
+# ============================================================================
+
+def generate_all_server_definitions_json(user_id):
+    """Generate JSON containing all server definitions for a user"""
+    _ensure_db_access()
+    servers = get_user_servers(user_id)
+
+    server_definitions = {}
+    for server in servers:
+        server_definitions[server.name] = server.definition
+
+    return json.dumps(server_definitions, indent=2, sort_keys=True)
+
+
+def store_server_definitions_cid(user_id):
+    """Store all server definitions as JSON in a CID and return the CID path"""
+    _ensure_db_access()
+    json_content = generate_all_server_definitions_json(user_id)
+    return store_cid_from_json(json_content, user_id)
+
+
+def get_current_server_definitions_cid(user_id):
+    """Get the CID path for the current server definitions JSON"""
+    _ensure_db_access()
+    json_content = generate_all_server_definitions_json(user_id)
+    json_bytes = json_content.encode('utf-8')
+    expected_cid = generate_cid(json_bytes)
+
+    existing_cid = get_cid_by_path(f"/{expected_cid}")
+    if existing_cid:
+        return expected_cid
+    return store_server_definitions_cid(user_id)
+
+
+def generate_all_variable_definitions_json(user_id):
+    """Generate JSON containing all variable definitions for a user"""
+    _ensure_db_access()
+    variables = get_user_variables(user_id)
+
+    variable_definitions = {}
+    for variable in variables:
+        variable_definitions[variable.name] = variable.definition
+
+    return json.dumps(variable_definitions, indent=2, sort_keys=True)
+
+
+def store_variable_definitions_cid(user_id):
+    """Store all variable definitions as JSON in a CID and return the CID path"""
+    _ensure_db_access()
+    json_content = generate_all_variable_definitions_json(user_id)
+    return store_cid_from_json(json_content, user_id)
+
+
+def get_current_variable_definitions_cid(user_id):
+    """Get the CID path for the current variable definitions JSON"""
+    _ensure_db_access()
+    json_content = generate_all_variable_definitions_json(user_id)
+    json_bytes = json_content.encode('utf-8')
+    expected_cid = generate_cid(json_bytes)
+
+    existing_cid = get_cid_by_path(f"/{expected_cid}")
+    if existing_cid:
+        return expected_cid
+    return store_variable_definitions_cid(user_id)
+
+
+def generate_all_secret_definitions_json(user_id):
+    """Generate JSON containing all secret definitions for a user"""
+    _ensure_db_access()
+    secrets = get_user_secrets(user_id)
+
+    secret_definitions = {}
+    for secret in secrets:
+        secret_definitions[secret.name] = secret.definition
+
+    return json.dumps(secret_definitions, indent=2, sort_keys=True)
+
+
+def store_secret_definitions_cid(user_id):
+    """Store all secret definitions as JSON in a CID and return the CID path"""
+    _ensure_db_access()
+    json_content = generate_all_secret_definitions_json(user_id)
+    json_bytes = json_content.encode('utf-8')
+    cid = generate_cid(json_bytes)
+
+    existing_cid = get_cid_by_path(f"/{cid}")
+    if not existing_cid:
+        create_cid_record(cid, json_bytes, user_id)
+
+    return cid
+
+
+def get_current_secret_definitions_cid(user_id):
+    """Get the CID path for the current secret definitions JSON"""
+    _ensure_db_access()
+    json_content = generate_all_secret_definitions_json(user_id)
+    json_bytes = json_content.encode('utf-8')
+    expected_cid = generate_cid(json_bytes)
+
+    existing_cid = get_cid_by_path(f"/{expected_cid}")
+    if existing_cid:
+        return expected_cid
+    return store_secret_definitions_cid(user_id)
+
+
+# ============================================================================
+# MIME TYPE AND CID SERVING HELPERS
+# ============================================================================
+
+EXTENSION_TO_MIME = {
+    'html': 'text/html',
+    'htm': 'text/html',
+    'txt': 'text/plain',
+    'css': 'text/css',
+    'js': 'application/javascript',
+    'json': 'application/json',
+    'xml': 'application/xml',
+    'pdf': 'application/pdf',
+    'zip': 'application/zip',
+    'tar': 'application/x-tar',
+    'gz': 'application/gzip',
+    'jpg': 'image/jpeg',
+    'jpeg': 'image/jpeg',
+    'png': 'image/png',
+    'gif': 'image/gif',
+    'svg': 'image/svg+xml',
+    'webp': 'image/webp',
+    'ico': 'image/x-icon',
+    'mp3': 'audio/mpeg',
+    'wav': 'audio/wav',
+    'ogg': 'audio/ogg',
+    'mp4': 'video/mp4',
+    'webm': 'video/webm',
+    'avi': 'video/x-msvideo',
+    'mov': 'video/quicktime',
+    'md': 'text/markdown',
+    'csv': 'text/csv',
+    'py': 'text/x-python',
+    'java': 'text/x-java-source',
+    'c': 'text/x-c',
+    'cpp': 'text/x-c++',
+    'h': 'text/x-c',
+    'hpp': 'text/x-c++',
+    'sh': 'application/x-sh',
+    'bat': 'application/x-msdos-program',
+    'exe': 'application/x-msdownload',
+    'dmg': 'application/x-apple-diskimage',
+    'deb': 'application/vnd.debian.binary-package',
+    'rpm': 'application/x-rpm'
+}
+
+MIME_TO_EXTENSION = {}
+for ext, mime in EXTENSION_TO_MIME.items():
+    if mime not in MIME_TO_EXTENSION:
+        MIME_TO_EXTENSION[mime] = ext
+
+
+def get_mime_type_from_extension(path):
+    """Determine MIME type from file extension in URL path"""
+    if '.' in path:
+        extension = path.split('.')[-1].lower()
+        return EXTENSION_TO_MIME.get(extension, 'application/octet-stream')
+    return 'application/octet-stream'
+
+
+def get_extension_from_mime_type(content_type):
+    """Get file extension from MIME type"""
+    base_mime = content_type.split(';')[0].strip().lower()
+    return MIME_TO_EXTENSION.get(base_mime, '')
+
+
+def extract_filename_from_cid_path(path):
+    """Extract filename from CID path for content disposition header."""
+    if path.startswith('/'):
+        path = path[1:]
+
+    if not path or path in ['.', '..']:
+        return None
+
+    parts = path.split('.')
+
+    if len(parts) < 3:
+        return None
+
+    filename_parts = parts[1:]
+    filename = '.'.join(filename_parts)
+
+    return filename
+
+
+def serve_cid_content(cid_content, path):
+    """Serve CID content with appropriate headers and caching"""
+    if cid_content is None or cid_content.file_data is None:
+        return None
+
+    cid = path[1:] if path.startswith('/') else path
+
+    content_type = get_mime_type_from_extension(path)
+
+    etag = f'"{cid.split(".")[0]}"'
+    if request.headers.get('If-None-Match') == etag:
+        response = make_response('', 304)
+        response.headers['ETag'] = etag
+        return response
+
+    if request.headers.get('If-Modified-Since'):
+        response = make_response('', 304)
+        response.headers['ETag'] = etag
+        response.headers['Last-Modified'] = cid_content.created_at.strftime('%a, %d %b %Y %H:%M:%S GMT')
+        return response
+
+    response = make_response(cid_content.file_data)
+    response.headers['Content-Type'] = content_type
+    response.headers['Content-Length'] = len(cid_content.file_data)
+
+    filename = extract_filename_from_cid_path(path)
+    if filename:
+        response.headers['Content-Disposition'] = f'attachment; filename="{filename}"'
+
+    response.headers['ETag'] = etag
+    response.headers['Last-Modified'] = cid_content.created_at.strftime('%a, %d %b %Y %H:%M:%S GMT')
+    response.headers['Cache-Control'] = 'public, max-age=31536000, immutable'
+    response.headers['Expires'] = 'Thu, 31 Dec 2037 23:55:55 GMT'
+
+    return response
