+++ routes.py
@@ -28,8 +28,22 @@ from db_access import (
 from forms import PaymentForm, TermsAcceptanceForm, FileUploadForm, InvitationForm, InvitationCodeForm, ServerForm, VariableForm, SecretForm
 from auth_providers import require_login, auth_manager, save_user_from_claims
 from secrets import token_urlsafe as secrets_token_urlsafe
-import hashlib
-import base64
+from cid_utils import (
+    generate_cid,
+    process_file_upload,
+    process_text_upload,
+    process_url_upload,
+    save_server_definition_as_cid,
+    store_server_definitions_cid,
+    get_current_server_definitions_cid,
+    store_variable_definitions_cid,
+    get_current_variable_definitions_cid,
+    store_secret_definitions_cid,
+    get_current_secret_definitions_cid,
+    get_mime_type_from_extension,
+    get_extension_from_mime_type,
+    serve_cid_content,
+)
 import traceback
 import json
 from text_function_runner import run_text_function
@@ -130,110 +144,6 @@ def privacy():
     """Display privacy policy"""
     return render_template('privacy.html')
 
-# ============================================================================
-# FILE UPLOAD AND CID HELPERS
-# ============================================================================
-
-def generate_cid(file_data):
-    """Generate a simple CID-like hash from file data only"""
-    # Create SHA-256 hash of the file content only
-    hasher = hashlib.sha256()
-    hasher.update(file_data)
-    sha256_hash = hasher.digest()
-
-    # Encode to base64 and create a CID-like string
-    encoded = base64.b32encode(sha256_hash).decode('ascii').lower().rstrip('=')
-    # Add CID prefix to make it look like an IPFS CID
-    return f"bafybei{encoded[:52]}"  # Truncate to reasonable length
-
-def process_file_upload(form):
-    """Process file upload from form and return file content and filename"""
-    uploaded_file = form.file.data
-    file_content = uploaded_file.read()
-    filename = uploaded_file.filename or 'upload'
-    return file_content, filename
-
-def process_text_upload(form):
-    """Process text upload from form and return file content"""
-    text_content = form.text_content.data
-    file_content = text_content.encode('utf-8')
-    return file_content
-
-def process_url_upload(form):
-    """Process URL upload from form by downloading content and return file content and MIME type"""
-    import requests
-    from urllib.parse import urlparse
-
-    url = form.url.data.strip()
-
-    try:
-        # Set reasonable timeout and headers
-        headers = {
-            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
-        }
-
-        response = requests.get(url, timeout=30, headers=headers, stream=True)
-        response.raise_for_status()
-
-        # Get MIME type from response headers
-        content_type = response.headers.get('content-type', 'application/octet-stream')
-        # Clean up MIME type (remove charset and other parameters)
-        mime_type = content_type.split(';')[0].strip().lower()
-
-        # Check content length to avoid downloading huge files
-        content_length = response.headers.get('content-length')
-        if content_length and int(content_length) > 100 * 1024 * 1024:  # 100MB limit
-            raise ValueError("File too large (>100MB)")
-
-        # Download content in chunks to handle large files efficiently
-        file_content = b''
-        downloaded_size = 0
-        max_size = 100 * 1024 * 1024  # 100MB limit
-
-        for chunk in response.iter_content(chunk_size=8192):
-            if chunk:
-                downloaded_size += len(chunk)
-                if downloaded_size > max_size:
-                    raise ValueError("File too large (>100MB)")
-                file_content += chunk
-
-        # Extract filename from URL for potential future use
-        parsed_url = urlparse(url)
-        filename = parsed_url.path.split('/')[-1]
-
-        # If no filename from URL or no extension, use MIME type to determine extension
-        if not filename or '.' not in filename:
-            extension = get_extension_from_mime_type(mime_type)
-            if extension:
-                filename = f"download{extension}"
-            else:
-                filename = "download"
-
-        return file_content, mime_type
-
-    except requests.exceptions.RequestException as e:
-        raise ValueError(f"Failed to download from URL: {str(e)}")
-    except Exception as e:
-        raise ValueError(f"Error processing URL: {str(e)}")
-def save_server_definition_as_cid(definition, user_id):
-    """Save server definition as CID and return the CID string"""
-    # Convert definition to bytes
-    definition_bytes = definition.encode('utf-8')
-
-    # Generate CID for the definition
-    cid = generate_cid(definition_bytes)
-
-    # Check if CID already exists to avoid duplicates
-    existing_cid = get_cid_by_path(f"/{cid}")
-    if not existing_cid:
-        create_cid_record(cid, definition_bytes, user_id)
-
-    return cid
-
-# ============================================================================
-# SERVER DEFINITIONS CID HELPERS
-# ============================================================================
-
 def get_server_definition_history(user_id, server_name):
     """Get historical server definitions for a specific server, ordered by time (latest first)"""
     from models import CID, db
@@ -296,52 +206,6 @@ def get_server_definition_history(user_id, server_name):
     
     return history
 
-def generate_all_server_definitions_json(user_id):
-    """Generate JSON containing all server definitions for a user"""
-    servers = get_user_servers(user_id)
-
-    # Create dictionary with server names as keys and definitions as values
-    server_definitions = {}
-    for server in servers:
-        server_definitions[server.name] = server.definition
-
-    # Convert to JSON string
-    return json.dumps(server_definitions, indent=2, sort_keys=True)
-
-def store_cid_from_json(json_content, user_id):
-    """Store all server definitions as JSON in a CID and return the CID path"""
-    json_bytes = json_content.encode('utf-8')
-
-    # Generate CID for the JSON content
-    cid = generate_cid(json_bytes)
-
-    # Check if CID already exists to avoid duplicates
-    existing_cid = get_cid_by_path(f"/{cid}")
-    if not existing_cid:
-        create_cid_record(cid, json_bytes, user_id)
-
-    return cid
-
-def store_server_definitions_cid(user_id):
-    """Store all server definitions as JSON in a CID and return the CID path"""
-    json_content = generate_all_server_definitions_json(user_id)
-    return store_cid_from_json(json_content, user_id)
-
-def get_current_server_definitions_cid(user_id):
-    """Get the CID path for the current server definitions JSON"""
-    # Generate what the current CID should be based on current servers
-    json_content = generate_all_server_definitions_json(user_id)
-    json_bytes = json_content.encode('utf-8')
-    expected_cid = generate_cid(json_bytes)
-
-    # Check if this CID exists in the database
-    existing_cid = get_cid_by_path(f"/{expected_cid}")
-    if existing_cid:
-        return expected_cid
-    else:
-        # If it doesn't exist, create it
-        return store_server_definitions_cid(user_id)
-
 def update_server_definitions_cid(user_id):
     """Update the server definitions CID after server changes"""
     return store_server_definitions_cid(user_id)
@@ -350,38 +214,6 @@ def update_server_definitions_cid(user_id):
 # VARIABLE DEFINITIONS CID HELPERS
 # ============================================================================
 
-def generate_all_variable_definitions_json(user_id):
-    """Generate JSON containing all variable definitions for a user"""
-    variables = get_user_variables(user_id)
-
-    # Create dictionary with variable names as keys and definitions as values
-    variable_definitions = {}
-    for variable in variables:
-        variable_definitions[variable.name] = variable.definition
-
-    # Convert to JSON string
-    return json.dumps(variable_definitions, indent=2, sort_keys=True)
-
-def store_variable_definitions_cid(user_id):
-    """Store all variable definitions as JSON in a CID and return the CID path"""
-    json_content = generate_all_variable_definitions_json(user_id)
-    return store_cid_from_json(json_content, user_id)
-
-def get_current_variable_definitions_cid(user_id):
-    """Get the CID path for the current variable definitions JSON"""
-    # Generate what the current CID should be based on current variables
-    json_content = generate_all_variable_definitions_json(user_id)
-    json_bytes = json_content.encode('utf-8')
-    expected_cid = generate_cid(json_bytes)
-
-    # Check if this CID exists in the database
-    existing_cid = get_cid_by_path(f"/{expected_cid}")
-    if existing_cid:
-        return expected_cid
-    else:
-        # If it doesn't exist, create it
-        return store_variable_definitions_cid(user_id)
-
 def update_variable_definitions_cid(user_id):
     """Update the variable definitions CID after variable changes"""
     return store_variable_definitions_cid(user_id)
@@ -389,48 +221,6 @@ def update_variable_definitions_cid(user_id):
 # ============================================================================
 # SERVER DEFINITIONS CID HELPERS
 # ============================================================================
-
-def generate_all_secret_definitions_json(user_id):
-    """Generate JSON containing all secret definitions for a user"""
-    secrets = get_user_secrets(user_id)
-
-    # Create dictionary with secret names as keys and definitions as values
-    secret_definitions = {}
-    for secret in secrets:
-        secret_definitions[secret.name] = secret.definition
-
-    # Convert to JSON string
-    return json.dumps(secret_definitions, indent=2, sort_keys=True)
-
-def store_secret_definitions_cid(user_id):
-    """Store all secret definitions as JSON in a CID and return the CID path"""
-    json_content = generate_all_secret_definitions_json(user_id)
-    json_bytes = json_content.encode('utf-8')
-
-    # Generate CID for the JSON content
-    cid = generate_cid(json_bytes)
-
-    # Check if CID already exists to avoid duplicates
-    existing_cid = get_cid_by_path(f"/{cid}")
-    if not existing_cid:
-        create_cid_record(cid, json_bytes, user_id)
-
-    return cid
-
-def get_current_secret_definitions_cid(user_id):
-    """Get the CID path for the current secret definitions JSON"""
-    # Generate what the current CID should be based on current secrets
-    json_content = generate_all_secret_definitions_json(user_id)
-    json_bytes = json_content.encode('utf-8')
-    expected_cid = generate_cid(json_bytes)
-
-    # Check if this CID exists in the database
-    existing_cid = get_cid_by_path(f"/{expected_cid}")
-    if existing_cid:
-        return expected_cid
-    else:
-        # If it doesn't exist, create it
-        return store_secret_definitions_cid(user_id)
 
 def update_secret_definitions_cid(user_id):
     """Update the secret definitions CID after secret changes"""
@@ -955,105 +745,6 @@ def execute_server_code(server, server_name):
         response.status_code = 500
         return response
 
-# ============================================================================
-# CID CONTENT SERVING HELPERS
-# ============================================================================
-
-# Shared MIME type and extension mappings
-EXTENSION_TO_MIME = {
-    'html': 'text/html',
-    'htm': 'text/html',
-    'txt': 'text/plain',
-    'css': 'text/css',
-    'js': 'application/javascript',
-    'json': 'application/json',
-    'xml': 'application/xml',
-    'pdf': 'application/pdf',
-    'zip': 'application/zip',
-    'tar': 'application/x-tar',
-    'gz': 'application/gzip',
-    'jpg': 'image/jpeg',
-    'jpeg': 'image/jpeg',
-    'png': 'image/png',
-    'gif': 'image/gif',
-    'svg': 'image/svg+xml',
-    'webp': 'image/webp',
-    'ico': 'image/x-icon',
-    'mp3': 'audio/mpeg',
-    'wav': 'audio/wav',
-    'ogg': 'audio/ogg',
-    'mp4': 'video/mp4',
-    'webm': 'video/webm',
-    'avi': 'video/x-msvideo',
-    'mov': 'video/quicktime',
-    'md': 'text/markdown',
-    'csv': 'text/csv',
-    'py': 'text/x-python',
-    'java': 'text/x-java-source',
-    'c': 'text/x-c',
-    'cpp': 'text/x-c++',
-    'h': 'text/x-c',
-    'hpp': 'text/x-c++',
-    'sh': 'application/x-sh',
-    'bat': 'application/x-msdos-program',
-    'exe': 'application/x-msdownload',
-    'dmg': 'application/x-apple-diskimage',
-    'deb': 'application/vnd.debian.binary-package',
-    'rpm': 'application/x-rpm'
-}
-
-# Create reverse mapping from MIME types to extensions
-MIME_TO_EXTENSION = {}
-for ext, mime in EXTENSION_TO_MIME.items():
-    if mime not in MIME_TO_EXTENSION:
-        # Use the first extension for each MIME type as the preferred one
-        MIME_TO_EXTENSION[mime] = ext
-
-def get_mime_type_from_extension(path):
-    """Determine MIME type from file extension in URL path"""
-    # Extract extension from path
-    if '.' in path:
-        extension = path.split('.')[-1].lower()
-        return EXTENSION_TO_MIME.get(extension, 'application/octet-stream')
-    # No extension, return default
-    return 'application/octet-stream'
-
-def get_extension_from_mime_type(content_type):
-    """Get file extension from MIME type"""
-    # Handle MIME types with parameters (e.g., "text/html; charset=utf-8")
-    base_mime = content_type.split(';')[0].strip().lower()
-    return MIME_TO_EXTENSION.get(base_mime, '')
-
-def extract_filename_from_cid_path(path):
-    """
-    Extract filename from CID path for content disposition header.
-
-    Rules:
-    - /{CID} -> no filename (no content disposition)
-    - /{CID}.{ext} -> no filename (no content disposition)
-    - /{CID}.{filename}.{ext} -> filename = {filename}.{ext}
-    """
-    # Remove leading slash
-    if path.startswith('/'):
-        path = path[1:]
-
-    # Handle empty or invalid paths
-    if not path or path in ['.', '..']:
-        return None
-    
-    # Split by dots
-    parts = path.split('.')
-
-    # Need at least 3 parts for filename: CID.filename.ext
-    if len(parts) < 3:
-        return None
-
-    # First part is CID, rest form the filename
-    filename_parts = parts[1:]
-    filename = '.'.join(filename_parts)
-
-    return filename
-
 def is_potential_versioned_server_path(path, existing_routes):
     """Check if path could be a versioned server invocation: /{server_name}/{partial_CID}"""
     if not path or not path.startswith('/'):
@@ -1154,52 +845,6 @@ def try_server_execution_with_partial(path):
     target = matches[0]
     definition_text = target.get('definition', '')
     return execute_server_code_from_definition(definition_text, server_name)
-
-def serve_cid_content(cid_content, path):
-    """Serve CID content with appropriate headers and caching"""
-    # Check if file_data is None (corrupted or missing data)
-    if cid_content is None or cid_content.file_data is None:
-        return None  # Return None to indicate content not found, let caller handle 404
-
-    # Extract CID from path (remove leading slash)
-    cid = path[1:] if path.startswith('/') else path
-
-    # Determine MIME type from URL extension
-    content_type = get_mime_type_from_extension(path)
-
-    # Check conditional requests using ETag (CID is perfect as ETag since it's content-based hash)
-    etag = f'"{cid.split(".")[0]}"'  # Use base CID without extension for ETag
-    if request.headers.get('If-None-Match') == etag:
-        # Client has the file cached, return 304 Not Modified
-        response = make_response('', 304)
-        response.headers['ETag'] = etag
-        return response
-
-    # Check If-Modified-Since (though ETag is more reliable for immutable content)
-    if request.headers.get('If-Modified-Since'):
-        # Since content is immutable, we can always return 304 if they have any cached version
-        response = make_response('', 304)
-        response.headers['ETag'] = etag
-        response.headers['Last-Modified'] = cid_content.created_at.strftime('%a, %d %b %Y %H:%M:%S GMT')
-        return response
-
-    # Serve the file with aggressive caching headers
-    response = make_response(cid_content.file_data)
-    response.headers['Content-Type'] = content_type
-    response.headers['Content-Length'] = len(cid_content.file_data)
-
-    # Set content disposition header if filename is indicated in the path
-    filename = extract_filename_from_cid_path(path)
-    if filename:
-        response.headers['Content-Disposition'] = f'attachment; filename="{filename}"'
-
-    # Immutable content caching headers
-    response.headers['ETag'] = etag
-    response.headers['Last-Modified'] = cid_content.created_at.strftime('%a, %d %b %Y %H:%M:%S GMT')
-    response.headers['Cache-Control'] = 'public, max-age=31536000, immutable'  # 1 year
-    response.headers['Expires'] = 'Thu, 31 Dec 2037 23:55:55 GMT'  # Far future
-
-    return response
 
 # ============================================================================
 # CRUD OPERATION HELPERS
