+++ test_server_cid_functionality.py
@@ -6,7 +6,7 @@ sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
 
 from app import app, db
 from models import User, Server, CID
-from routes import save_server_definition_as_cid
+from cid_utils import save_server_definition_as_cid
 
 def test_server_cid_functionality():
     """Test that server definitions are saved as CIDs when created/updated"""
